+++ components/brave_news/browser/BUILD.gn
@@ -30,6 +30,8 @@ static_library("browser") {
     "feed_controller.h",
     "feed_fetcher.cc",
     "feed_fetcher.h",
+    "feed_sampling.cc",
+    "feed_sampling.h",
     "feed_v2_builder.cc",
     "feed_v2_builder.h",
     "html_parsing.cc",
