+++ components/brave_news/browser/channels_controller.h
@@ -7,6 +7,7 @@
 #define BRAVE_COMPONENTS_BRAVE_NEWS_BROWSER_CHANNELS_CONTROLLER_H_
 
 #include <string>
+#include <vector>
 
 #include "base/containers/flat_map.h"
 #include "base/functional/callback_forward.h"
@@ -26,6 +27,11 @@ using ChannelsCallback = base::OnceCallback<void(Channels)>;
 
 inline constexpr char kTopSourcesChannel[] = "Top Sources";
 inline constexpr char kTopNewsChannel[] = "Top News";
+
+// Get all the channels for a publisher in a specific locale.
+std::vector<std::string> GetChannelsForPublisher(
+    const std::string& locale,
+    const mojom::PublisherPtr& publisher);
 
 class ChannelsController {
  public:
